[wnsxor1993#28] chore: 파일정리
