+++ iOS/IssueTracker/IssueTracker/Network/URLComponent/Protocol/URLConfigurable.swift
@@ -1,0 +1,14 @@
+//
+//  URLConfigurable.swift
+//  IssueTracker
+//
+//  Created by Kai Kim on 2022/06/16.
+//
+
+import Foundation
+
+protocol URLConfigurable {
+    var host: String {get}
+    var path: String {get}
+}
+
